+++ ventiq_app/lib/services/web_summary_printer_service_stub.dart
@@ -6,7 +6,7 @@ class WebSummaryPrinterServiceImpl {
   Future<bool> printDetailedSummary({
     required List<OrderItem> productosVendidos,
     required double totalVentas,
-    required int totalProductos,
+    required double totalProductos,
     required double totalEgresado,
     required double totalEfectivoReal,
   }) async {
